+++ tempest_fastapi_sdk/auth/__init__.py
@@ -34,6 +34,7 @@ from tempest_fastapi_sdk.auth.schemas import (
     ActivationResponseSchema as ActivationResponseSchema,
 )
 from tempest_fastapi_sdk.auth.schemas import ActivationToken as ActivationToken
+from tempest_fastapi_sdk.auth.schemas import AuthUserSchema as AuthUserSchema
 from tempest_fastapi_sdk.auth.schemas import (
     EmailChangeConfirmSchema as EmailChangeConfirmSchema,
 )
@@ -97,6 +98,7 @@ __all__: list[str] = [
     "ActivationResponseSchema",
     "ActivationToken",
     "AuthCookieConfig",
+    "AuthUserSchema",
     "EmailChangeConfirmSchema",
     "EmailChangeRequestSchema",
     "EmailChangeResponseSchema",